Fuck the meta

+++ Garbage/Assets/Completed/Scripts/PlayerVariables.cs
@@ -5,8 +5,8 @@ public class PlayerVariables : MonoBehaviour
 {
     public int maxHealth;
     public int currentHealth;
-    public int horizontal_speed;
-    public int vertical_speed;
+    public float horizontal_speed;
+    public float vertical_speed;
     public bool immune;
     public bool knockdown;
     public bool immobile;
